+++ Sources/VCore/Models/CastingError.swift
@@ -11,6 +11,7 @@ import Foundation
 /// Error that occurs during casting.
 @MemberwiseInitializable(accessLevelModifier: .private)
 public struct CastingError: BaseErrorProtocol, Sendable {
+    // MARK: Properties
     public static let domain: String = "com.vcore"
     public let code: Int
     public let description: String