+++ Sources/VCore/Models/Layout Directions/LayoutDirection+Helpers.swift
@@ -1,5 +1,5 @@
 //
-//  LayoutDirectionHorizontal.swift
+//  LayoutDirection.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 27.02.23.
@@ -7,24 +7,14 @@
 
 import SwiftUI
 
-// MARK: - Layout Direction Horizontal
-/// Horizontal directional layout in which content can be laid out.
-///
-/// Unlike `SwiftUI.LayoutDirection`, `LayoutDirectionHorizontal` is exhaustive.
-public enum LayoutDirectionHorizontal: Int, Sendable, CaseIterable {
-    // MARK: Cases
-    /// Left-to-right direction.
-    case leftToRight
-    
-    /// Right-to-left direction.
-    case rightToLeft
-    
-    // MARK: Properties
+// MARK: - Layout Direction
+extension LayoutDirection {
     /// Indicates if layout direction is reversed.
     public var isReversed: Bool {
         switch self {
         case .leftToRight: false
         case .rightToLeft: true
+        @unknown default: false
         }
     }
 
@@ -34,6 +24,7 @@ public enum LayoutDirectionHorizontal: Int, Sendable, CaseIterable {
         switch self {
         case .leftToRight: .rightToLeft
         case .rightToLeft: .leftToRight
+        @unknown default: .rightToLeft
         }
     }
 
@@ -43,6 +34,7 @@ public enum LayoutDirectionHorizontal: Int, Sendable, CaseIterable {
         switch self {
         case .leftToRight: .leading
         case .rightToLeft: .trailing
+        @unknown default: .leading
         }
     }
     
@@ -51,6 +43,7 @@ public enum LayoutDirectionHorizontal: Int, Sendable, CaseIterable {
         switch self {
         case .leftToRight: .leading
         case .rightToLeft: .trailing
+        @unknown default: .leading
         }
     }
     
@@ -59,6 +52,7 @@ public enum LayoutDirectionHorizontal: Int, Sendable, CaseIterable {
         switch self {
         case .leftToRight: .leading
         case .rightToLeft: .trailing
+        @unknown default: .leading
         }
     }
 }